Move feature/report/components/option to common components
