+++ app/javascript/mastodon/components/option.jsx
