+++ app/javascript/mastodon/features/report/category.jsx
@@ -8,8 +8,7 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import { connect } from 'react-redux';
 
 import { Button } from 'mastodon/components/button';
-
-import Option from './components/option';
+import Option from 'mastodon/components/option';
 
 const messages = defineMessages({
   dislike: { id: 'report.reasons.dislike', defaultMessage: 'I don\'t like it' },
