+++ app/javascript/mastodon/features/report/components/status_check_box.jsx
@@ -6,11 +6,10 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import { Avatar } from 'mastodon/components/avatar';
 import { DisplayName } from 'mastodon/components/display_name';
 import MediaAttachments from 'mastodon/components/media_attachments';
+import Option from 'mastodon/components/option';
 import { RelativeTimestamp } from 'mastodon/components/relative_timestamp';
 import StatusContent from 'mastodon/components/status_content';
 import { VisibilityIcon } from 'mastodon/components/visibility_icon';
-
-import Option from './option';
 
 class StatusCheckBox extends PureComponent {
